backend/requirements.txt
cryptography==38.0.1
dj-database-url==0.5.0
Django==3.2
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
dnspython==2.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.26.0
rsa==4.9
six==1.16.0
sqlparse==0.4.2
static==1.1.1
urllib3==1.26.7
